+++ client/main.js
@@ -15,5 +15,5 @@ import '/imports/channel/client';
 import '/imports/roster/client';
 import '/imports/charts/client';
 import '/imports/search/client';
-import '/imports/shares/client';
+import '/imports/shared/client';
 import '/imports/cordova/client';