+++ imports/model/instance.js
@@ -1,6 +1,7 @@
 // - company
 // - domain
 // - adminId
+// - sharedId
 Instance = new Mongo.Collection('instance');
 
 Instance.get = () => {