+++ imports/shared/shared.js
@@ -1,4 +1,4 @@
-ThreadCategories.add("Shares", {
+ThreadCategories.add("Shared", {
   icon: "fa fa-connectdevelop fa-1-2x",
   iconUnread: "fa fa-connectdevelop fa-1-2x",
   description: "TODO: ",