+++ server/main.js
@@ -13,5 +13,5 @@ import '/imports/channel/server';
 import '/imports/roster/server';
 import '/imports/charts/server';
 import '/imports/search/server';
-import '/imports/shares/server';
+import '/imports/shared/server';
 import '/imports/cordova/server';